GatherRate fix

+++ Pluton/World.cs
@@ -10,7 +10,7 @@ namespace Pluton
 
         private static World instance;
 
-        public double ResourceGatherMultiplier = 1.0d;
+        public float ResourceGatherMultiplier = 1.0f;
         public Timer freezeTimeTimer;
         private float frozenTime = -1;
